Revert "bakery pos components restructured and products search  added"
