+++ src/lang/bn/translation.json
@@ -157,7 +157,7 @@
   "ManageProduct": "পণ্য পরিচালনা",
   "Unit": "ইউনিট",
   "Brand": "ব্র্যান্ড",
-  "PurchasePrice": "ক্রয় মূল্য",
+  "PurchasePrice": "পার্চেজ প্রাইস",
   "CreateProduct": "প্রোডাক্ট তৈরী",
   "ProductInformation": "পণ্যের বিবরণ",
   "AlternativeProductName": "পণ্যের বিকল্প নাম",
@@ -457,7 +457,7 @@
   "UnitMeasurement" : "ইউনিট মেজারমেন্ট",
   "BranchManagement" : "ব্রাঞ্চ ম্যানেজমেন্ট",
   "VatManagement" : "ভ্যাট ব্যবস্থাপনা",
-  "SkuItem" : "SKU আইটেম",
+  "SKUItem" : "SKU আইটেম",
   "AddProvision" : "অ্যাড প্রভিশন",
   "StockDetails" : "স্টক ডিটেইলস",
   "StockOverview" : "স্টক ওভারভিউ",
@@ -470,16 +470,5 @@
   "SettingsDetailsData" : "সেটিংসের বিস্তারিত",
   "NoUppercaseAllowedMessage" : "বড় হাতের অক্ষর গ্রহণযোগ্য নয়",
   "ManageStock" : "স্টক ম্যানেজ",
-  "ManageProductionCost" : "উৎপাদন খরচ পরিচালনা",
-  "DisplayName" : "ডিসপ্লে নাম",
-  "LanguageName" : "ল্যাঙ্গুয়েজ নাম",
-  "Description" : "বর্ণনা",
-  "RawMaterials" : "কাঁচামাল",
-  "Grade" : "গ্রেড",
-  "BatchInvoice" : "ব্যাচ ইনভয়েস",
-  "Provision" : "প্রভিশন",
-  "Stockable" : "স্টকেবল",
-  "PostProduction" : "পোস্ট প্রোডাকশন",
-  "MidProduction" : "মিড প্রোডাকশন",
-  "PreProduction" : "প্রি প্রোডাকশন"
+  "ManageProductionCost" : "উৎপাদন খরচ পরিচালনা"
 }
