+++ src/lang/en/translation.json
@@ -619,6 +619,7 @@
   "AlternativeEmail" : "Alternative Email",
   "ChooseProductGrade" : "Choose Product Grade",
   "VatManagement" : "Vat Management",
+  "SKUItem" : "SKU Item",
   "ChooseHS" : "Choose HS",
   "SkuItem" : "Sku Item",
   "EnterQuantity" : "Enter Quantity",
@@ -697,10 +698,5 @@
   "EnterOpeningQuantity" : "Enter opening quantity",
   "EnterPurchasePrice" : "Enter purchase price",
   "SalesProcess" : "Sales process",
-  "CustomerMobileNumber" : "Customer Mobile Number",
-  "ValidationError" : "Validation Error",
-  "Description" : "Description",
-  "Grade" : "Grade",
-  "Provision" : "Provision",
   "": ""
 }
